モーフ係数のlerpで整数計算していて小数点以下が切り捨てられていたのを修正

+++ pkg/vmd/morph_frame.go
@@ -40,8 +40,8 @@ func (nextMf *MorphFrame) lerpFrame(prevFrame IBaseFrame, index int) IBaseFrame 
 
 	mf := NewMorphFrame(index)
 
-	ry := (index - prevIndex) / (nextIndex - prevIndex)
-	mf.Ratio = prevMf.Ratio + (nextMf.Ratio-prevMf.Ratio)*float64(ry)
+	ry := float64(index-prevIndex) / float64(nextIndex-prevIndex)
+	mf.Ratio = prevMf.Ratio + (nextMf.Ratio-prevMf.Ratio)*ry
 
 	return mf
 }
